+++ NEWS.md
@@ -6,6 +6,8 @@
   - Options now included CV for the model residuals, mean-squared error (MSE), or for the standard error of measurement (SEM)
 - Updated plots to include the `delta` argument on plot if specified, and estimate of LoA
 - Added agreement coefficient function (`agree_coef`).
+- Assumptions checks are added for all "simple-agree" class results
+  - Checks include: normality, heteroscedasticity, and proportional bias
 
 ## v0.0.3
 - Fixed error in `agree_nest` and `agree_reps`; now properly handles missing values